+++ mapchete_hub/commands/index.py
@@ -1,22 +1,33 @@
-from celery.utils.log import get_task_logger
+import logging
 from mapchete.io import makedirs
 import os
+from shapely import wkt
 
-from mapchete_hub.commands._misc import announce_on_slack
+from mapchete_hub.commands._misc import send_slack_message
 from mapchete_hub.celery_app import celery_app
 from mapchete_hub.commands._index import mapchete_index
 
 
-logger = get_task_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 # bind=True enables getting the job ID and sending status updates (with send_events())
 # ignore_result=True important, otherwise it will be stored in broker
 @celery_app.task(bind=True, ignore_result=True)
-def run(self, *args, mapchete_config=None, process_area=None, **kwargs):
+def run(
+    self,
+    *args,
+    mapchete_config=None,
+    process_area=None,
+    announce_on_slack=False,
+    **kwargs
+):
     """Celery task for mapchete_index."""
     logger.info("got job %s", self.request.id)
-    logger.debug("initialize process")
+    logger.debug("extra kwargs: %s", kwargs)
+    process_area = wkt.loads(process_area)
+
+    # send first event in order to have an empty progress_data dictionary
     self.send_event('task-progress', progress_data=dict(current=None, total=None))
 
     # first, the inputs get parsed, i.e. all metadat queried from catalogue
@@ -33,6 +44,8 @@ def run(self, *args, mapchete_config=None, process_area=None, **kwargs):
         makedirs(index_output_path)
     else:
         index_output_path = mapchete_config["output"]["path"]
+
+    logger.info("preparing execute process")
     executor = mapchete_index(
         mapchete_config=mapchete_config,
         process_area=process_area,
@@ -44,7 +57,7 @@ def run(self, *args, mapchete_config=None, process_area=None, **kwargs):
     # get total tiles
     total_tiles = next(executor)
     self.send_event('task-progress', progress_data=dict(current=0, total=total_tiles))
-    logger.debug("total tiles: %s", total_tiles)
+    logger.info("indexing from %s process tiles", total_tiles)
 
     # iterate over finished process tiles and update task state
     for i, _ in enumerate(executor):
@@ -53,4 +66,5 @@ def run(self, *args, mapchete_config=None, process_area=None, **kwargs):
         self.send_event('task-progress', progress_data=dict(current=i, total=total_tiles))
 
     logger.info("processing successful.")
-    announce_on_slack(mapchete_config=mapchete_config, process_area=process_area)
+    if announce_on_slack:
+        send_slack_message(process_area.centroid.x, process_area.centroid.y)